seperated logic to a seperate controller module with respective methods to handle endpoints related to plants
